requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cookies==2.2.1
coverage==4.0.3
idna==2.6
pep8==1.7.0
py==1.4.34
pytest==3.2.1
pytest-cov==2.5.1
python-coveralls==2.9.1
python-dateutil==2.6.1
PyYAML==3.12
requests==2.18.4
responses==0.7.0
six==1.10.0
tabulate==0.7.7
urllib3==1.22
